Refactor builder logic into dedicated modules
